Extend SplashActivity to launch any activity.
If authenticator_activity is started without launching the
main activity, neko will not be initialized and on-ui throws
an exception. So instead of starting authenticator_activity
directly, ask SplashActivity to do it for us and initialize
neko if it wasn't already.
I'm probably abusing the Intent selector, but it seems to work.